changed the way singleton instances are registered to the event bus
we now explicitly register all the event subscribers to the event bus through a method call
instead of relying on Guice's singleton pre-loading behavior. This makes the singleton registration
more testable. the classes that needs to be registered with the eventbus are collected using
the multibinding plugin for Guice. each module contributes classes that needs to be registered
with the eventbus by adding them to a multibinder for the EventSubscriber interface.