Fixed a bug where we are adding Remove component to wrong entity after a new level is loaded.
Fixed by not storing references to entity outside of engine.
Character activation effects is now handled by CharacterSystem.
Correctness + easy implementation > small inefficiency
The inefficiency is actually not even that bad. We are not dealing with
hundreds of characters and we are not looping over them every frame.
Looping over them ones when an user prompted action happens is not
that costly.