added new turn class with the concept of a stage
We now save the specific stage the world state is in (i.e. start, player,
end, or enemy). The appropriate stage is then loaded when the game
resumes. This will help keep the state of the game always in sync
with the script no matter when the player quit the game. This also
make it easier to describe the different script triggers since now
we can explicitly target the stage of a turn. The states now also
check they are in the correct stage before proceeding.